requirements.txt
mysql-connector-python
pandas
Pillow
reportlab
tkcalendar
python-dotenv